Change guards for importing mined blocks to allow mining on single-node private networks
